==> notebooks/01_cleanup_CBTTC_manifest.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"></ul></div>

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

Cleanup rsem manifest

CBTTC_RNA_manifest_genes <- read.csv(paste0(manifestpath, "CBTTC_RSEM-manifest.csv"),
                               header = TRUE, stringsAsFactors = FALSE)

CBTTC_celllines <- read.csv(paste0(manifestpath,"cell_line_RNAalqids_CBTTC.csv"), 
                             header = TRUE, stringsAsFactors = F)

CBTTC_celllines$X <- NULL

CBTTC_celllines

CBTTC_RNA_manifest_genes_linerm <- CBTTC_RNA_manifest_genes[!CBTTC_RNA_manifest_genes$aliquot_id %in% 
                                                            CBTTC_celllines$cell_lines_rnaalq_id,]

dim(CBTTC_RNA_manifest_genes)
dim(CBTTC_RNA_manifest_genes_linerm)

CBTTC_RNA_manifest_genes_linerm_dedup <- CBTTC_RNA_manifest_genes_linerm[!duplicated(CBTTC_RNA_manifest_genes_linerm$sample_id),]

dim(CBTTC_RNA_manifest_genes_linerm)
dim(CBTTC_RNA_manifest_genes_linerm_dedup)

table(CBTTC_RNA_manifest_genes_linerm_dedup$Tumor.Descriptor)

Manually checked all aliquot ids of duplicated samples. they are all tumor rna aliquots

write.csv(CBTTC_RNA_manifest_genes_linerm_dedup,
         file = paste0(manifestpath, "CBTTC_RSEM-manifest_linerm_dedup.csv"), row.names = F)

clean up mixcr manifest

mixcr_manifest <- read.csv(paste0(manifestpath, "mixcr_manifest.csv"),
                               header = TRUE, stringsAsFactors = FALSE)

separate cbttc from the rest

cbttc_mixcr_manifest <- mixcr_manifest[ grepl("CLONES_TRB_7316",mixcr_manifest$name),]
other_mixcr_manifest <- mixcr_manifest[!grepl("CLONES_TRB_7316",mixcr_manifest$name),]

cbttc_mixcr_manifest_linerm <- cbttc_mixcr_manifest[!cbttc_mixcr_manifest$aliquot_id %in% 
                                       CBTTC_celllines$cell_lines_rnaalq_id,]

dim(cbttc_mixcr_manifest)
dim(cbttc_mixcr_manifest_linerm)

cbttc_mixcr_manifest_linerm_dedup <- cbttc_mixcr_manifest_linerm[!duplicated(cbttc_mixcr_manifest_linerm$sample_id),]

dim(cbttc_mixcr_manifest_linerm_dedup)

bind all back together and save

mixcr_manifest_cleaned <- rbind(cbttc_mixcr_manifest_linerm_dedup,
                               other_mixcr_manifest)

write.csv(mixcr_manifest_cleaned,
         file = paste0(manifestpath, "mixcr_manifest_cleaned.csv"), row.names = F)
