==> notebooks/03_exp_matrix_4ESTIMATE.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Functions" data-toc-modified-id="Functions-2">Functions</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-3">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-4">Main</a></span><ul class="toc-item"><li><span><a href="#CBTTC" data-toc-modified-id="CBTTC-4.1">CBTTC</a></span></li></ul></li></ul></div>

# Dependencies

library(tidyr)

# Functions

matchexp_matrixfx <- function(exp_matrix, estimate_manifest, group, sample_type, matchingcol){
    
    message("dimensions of exp_matrix: ", deparse(substitute(exp_matrix)))
    print(dim(exp_matrix))
    
    subset_estimate_manifest <- estimate_manifest[estimate_manifest$group == group,]
    subset_estimate_manifest_sampletype <- subset_estimate_manifest[grepl(sample_type,
                                                                    subset_estimate_manifest$sample_type),]
    
    message("dimensions of estimate_manifest_df for: ", group, " sample_type:", sample_type)
    print(dim(subset_estimate_manifest_sampletype))
    
    exp_matrix_matchingset <- exp_matrix[,colnames(exp_matrix) %in% 
                                         subset_estimate_manifest_sampletype[[matchingcol]]]
    
    message("dimensions of exp_matrix_matchingset for: ", deparse(substitute(exp_matrix)), 
            " with matching column: ", matchingcol)
    print(dim(exp_matrix_matchingset))
    
    exp_matrix_matchingset <- cbind(exp_matrix[,1:2], exp_matrix_matchingset)

    return(exp_matrix_matchingset)
}

longtowidefx <- function(long_exp_matrix){ 
    wide_exp_matrix <- spread(long_exp_matrix, key = "sample", value = "tpm")
    return(wide_exp_matrix)
}

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

## CBTTC

From Cavatica

CBTTCtpm <- read.table(paste0(datapath, "exp_mat/KF_CBTTC_tpm_rsem_matrix.txt_sampleid.txt"), sep = "\t",
                    header = TRUE, stringsAsFactors = FALSE, check.names = F)

head(CBTTCtpm)

CBTTCtpm <- cbind(NA, NA, CBTTCtpm)
colnames(CBTTCtpm)[1:2] <- c("ensembl_id", "hgnc_symbol")

CBTTCtpm$hgnc_symbol <- gsub(".*_", "", CBTTCtpm$`Var.3`)
CBTTCtpm$ensembl_id <- gsub("_.*", "", CBTTCtpm$`Var.3`)
CBTTCtpm$`Var.3` <- NULL

dim(CBTTCtpm)

CBTTCtpm <- CBTTCtpm[!duplicated(CBTTCtpm$hgnc_symbol),]

dim(CBTTCtpm)

head(CBTTCtpm)

CBTTCtpm[ CBTTCtpm$hgnc_symbol == "DAZ1", "7316-905"]

write.table(CBTTCtpm,
            file = paste0(datapath, "exp_mat/CBTTC_tpm_matrix_dedup.txt"),
                sep = "\t", quote = F, row.names = F)

write.table(CBTTCtpm[,colnames(CBTTCtpm) !="ensembl_id"],
            file = paste0(datapath, "exp_mat/CBTTC_tpm_matrix_dedup_hgnc.txt"),
                sep = "\t", quote = F, row.names = F)
