==> notebooks/04_ESTIMATE_dataframe.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Functions" data-toc-modified-id="Functions-2">Functions</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-3">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-4">Main</a></span><ul class="toc-item"><li><span><a href="#Load-and-cleanup-estimate-files" data-toc-modified-id="Load-and-cleanup-estimate-files-4.1">Load and cleanup estimate files</a></span><ul class="toc-item"><li><span><a href="#GDC" data-toc-modified-id="GDC-4.1.1">GDC</a></span></li><li><span><a href="#DKFZ" data-toc-modified-id="DKFZ-4.1.2">DKFZ</a></span></li><li><span><a href="#CBTTC" data-toc-modified-id="CBTTC-4.1.3">CBTTC</a></span></li></ul></li><li><span><a href="#Create-an-estimate_manifest-file" data-toc-modified-id="Create-an-estimate_manifest-file-4.2">Create an estimate_manifest file</a></span><ul class="toc-item"><li><span><a href="#merge-estimate-and-manifest-of-KF-and-DKFZ-by-sample_iq" data-toc-modified-id="merge-estimate-and-manifest-of-KF-and-DKFZ-by-sample_iq-4.2.1">merge estimate and manifest of KF and DKFZ by sample_iq</a></span></li><li><span><a href="#merge-estimate-and-manifest-of-GDC-by-aliquot_iq" data-toc-modified-id="merge-estimate-and-manifest-of-GDC-by-aliquot_iq-4.2.2">merge estimate and manifest of GDC by aliquot_iq</a></span></li></ul></li><li><span><a href="#Subset-to-primary-tumours" data-toc-modified-id="Subset-to-primary-tumours-4.3">Subset to primary tumours</a></span></li></ul></li></ul></div>

# Dependencies

library(dplyr)
library(reshape2)
library(gtable)

# Functions 


to replace ids with barcodes using the map file (Provided by Komal)

id2barcodefx <- function(inpath,estimate_outputname){
    estimate_output <- read.table(paste0(inpath, estimate_outputname), 
                              sep = "\t", header = TRUE, stringsAsFactors = FALSE)
    message("Currnet estimate file:")
    print(estimate_outputname)
    message("dimensions of estimate output file:")
    print(dim(estimate_output))
    estimate_output$SampleID <- as.character(estimate_output$SampleID)
# remove the X from sampleid
    estimate_output$SampleID <- gsub("X", "", estimate_output$SampleID)
# replace "." with "-"
    estimate_output$SampleID <- gsub("[.]", "-", estimate_output$SampleID)
    non_existing <- estimate_output[,1][!estimate_output[,1] %in% mappings$sample_id]
    non_existing <- as.character(non_existing)
    message("cannot find this many in mappings")
    print(length(non_existing))
    exp_matrix_subset <- estimate_output[!estimate_output[,1] %in% non_existing,]
    mappings$barcodes <- as.character(mappings$barcodes)
    exp_matrix_subset[,1] <- mappings$barcodes[match(exp_matrix_subset[,1],mappings$sample_id)]
    message("Dim after id2barcode replacement")
    print(dim(exp_matrix_subset))
    return(exp_matrix_subset)
}

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

load(paste0(manifestpath,"Master_manifest.RData"))

table(IPD_metadata$group)

## Load and cleanup estimate files

### GDC

Change TCGA and TARGET ids to barcodes

mappings <- read.table(paste0(manifestpath,"sampleid_barcode_mappings.txt"),
                       sep = "\t", header = T, stringsAsFactors = FALSE)

mappings$barcodes <- as.character(mappings$sample_barcode)

Got the output from pengbo

NBL_estimate <- id2barcodefx(paste(datapath,"ESTIMATE/ESTIMATE_output/", sep = ""),
                             "TARGET_NBL_ESTIMATEoutput.txt")
COAD_estimate <- id2barcodefx(paste(datapath,"ESTIMATE/ESTIMATE_output/", sep = ""),
                             "TCGA_COAD_ESTIMATEoutput.txt")
LUAD_estimate <- id2barcodefx(paste(datapath,"ESTIMATE/ESTIMATE_output/", sep = ""),
                             "TCGA_LUAD_ESTIMATEoutput.txt")
SKCM_estimate <- id2barcodefx(paste(datapath,"ESTIMATE/ESTIMATE_output/", sep = ""),
                             "TCGA_SKCM_ESTIMATEoutput.txt")
PRAD_estimate <- id2barcodefx(paste(datapath,"ESTIMATE/ESTIMATE_output/", sep = ""),
                             "TCGA_PRAD_ESTIMATEoutput.txt")
GBM_estimate <- id2barcodefx(paste(datapath,"ESTIMATE/ESTIMATE_output/", sep = ""),
                             "TCGA_GBM_ESTIMATEoutput.txt")
LGG_estimate <- id2barcodefx(paste(datapath,"ESTIMATE/ESTIMATE_output/", sep = ""),
                             "TCGA_LGG_ESTIMATEoutput.txt")
OV_estimate <- id2barcodefx(paste(datapath,"ESTIMATE/ESTIMATE_output/", sep = ""),
                             "TCGA_OV_ESTIMATEoutput.txt")

### DKFZ

ATRT_estimate <- read.table(paste0(datapath, "ESTIMATE/ESTIMATE_output/ICGC_ATRTs.normalizedTPMs_perGencodeV26Gene.sleuth.estimateOutput.txt"),
                           sep = "\t", header = TRUE, stringsAsFactors = FALSE)

ET_estimate <- read.table(paste0(datapath, "ESTIMATE/ESTIMATE_output/ICGC_ETs.normalizedTPMs_perGencodeV26Gene.sleuth.estimateOutput.txt"),
                           sep = "\t", header = TRUE, stringsAsFactors = FALSE)

MB_estimate <- read.table(paste0(datapath,"ESTIMATE/ESTIMATE_output/ICGC_MBs.normalizedTPMs_perGencodeV26Gene.sleuth.estimateOutput.txt"),
                           sep = "\t", header = TRUE, stringsAsFactors = FALSE)

PA_estimate <- read.table(paste0(datapath,"ESTIMATE/ESTIMATE_output/ICGC_PAs.normalizedTPMs_perGencodeV26Gene.sleuth.estimateOutput.txt"),
                           sep = "\t", header = TRUE, stringsAsFactors = FALSE)

pedGBM_estimate <- read.table(paste0(datapath, "ESTIMATE/ESTIMATE_output/ICGC_GBMs.normalizedTPMs_perGencodeV26Gene.sleuth.estimateOutput.txt"),
                           sep = "\t", header = TRUE, stringsAsFactors = FALSE)

PDX_estimate <- read.table(paste0(datapath, "ESTIMATE/ESTIMATE_output/PDX_2017Oct06.normalizedTPMs_perGencodeV26Gene.sleuth.estimateOutput.txt"),
                           sep = "\t", header = TRUE, stringsAsFactors = FALSE)

ATRT_estimate$SampleID <- gsub("[.]", "-", ATRT_estimate$SampleID)

### CBTTC


CBTTC_estimate <- read.table(paste0(datapath, "ESTIMATE/ESTIMATE_output/ESTIMATEOutput.CBTTC.txt"),
                              sep = "\t", header = TRUE, check.names = F,
                            stringsAsFactors =F, row.names = NULL)

dim(CBTTC_estimate)

head(CBTTC_estimate)

colnames(CBTTC_estimate)[colnames(CBTTC_estimate) == "row.names"] <- "SampleID"

CBTTC_estimate$SampleID <- gsub("X", "",CBTTC_estimate$SampleID)
CBTTC_estimate$SampleID <- gsub("7316.", "7316-",CBTTC_estimate$SampleID)

head(CBTTC_estimate)

## Create an estimate_manifest file

KF_DKFZ_estimate <- rbind(PA_estimate, #DKFZ
                        ET_estimate,
                        pedGBM_estimate,
                        ATRT_estimate,
                        MB_estimate,
                        CBTTC_estimate )

### merge estimate and manifest of KF and DKFZ by sample_iq

colnames(KF_DKFZ_estimate)[colnames(KF_DKFZ_estimate) == "SampleID"] <- "sample_id"

KF_DKFZ_estimate_manifest <- merge(KF_DKFZ_estimate, IPD_metadata, by = "sample_id")

table(KF_DKFZ_estimate_manifest$group)

### merge estimate and manifest of GDC by aliquot_iq

GDC_estimate <- rbind(NBL_estimate,
                     OV_estimate,
                     PRAD_estimate,
                     COAD_estimate,
                     SKCM_estimate,
                     GBM_estimate,
                     LGG_estimate,
                     LUAD_estimate)

colnames(GDC_estimate)[colnames(GDC_estimate) == "SampleID"] <- "aliquot_id"

GDC_estimate_manifest <- merge(GDC_estimate, IPD_metadata, by = "aliquot_id")

17 TCGA samples missing from IPD_metadata, because their raw files were not available through cavatica, but their tpm file was forwarded to me by komal

Merge both

total_estimate_manifest <- rbind(GDC_estimate_manifest,
                                KF_DKFZ_estimate_manifest)

## Subset to primary tumours 


estimate_manifest_primary <- total_estimate_manifest[grepl("Initial",total_estimate_manifest$sample_type) |
                                                     grepl("Primary", total_estimate_manifest$sample_type), ]

estimate_manifest_primary <- rbind(estimate_manifest_primary,
                                  total_estimate_manifest[total_estimate_manifest$group == "KF_NBL",])

dim(estimate_manifest_primary)

table(total_estimate_manifest$group)

table(estimate_manifest_primary$group)

save(estimate_manifest_primary,
     file = paste0(datapath, "ESTIMATE/estimate_manifest_primary.RData"))   

save(PDX_estimate,
     file = paste0(datapath, "ESTIMATE/PDX_estimate.RData"))
