==> notebooks/05_Cohort_cleanup.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-2">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-3">Main</a></span><ul class="toc-item"><li><span><a href="#Cleanup-according-to-CRC-review" data-toc-modified-id="Cleanup-according-to-CRC-review-3.1">Cleanup according to CRC review</a></span></li><li><span><a href="#Modify-disease-types-based-on-my-path-reviews" data-toc-modified-id="Modify-disease-types-based-on-my-path-reviews-3.2">Modify disease types based on my path reviews</a></span></li><li><span><a href="#Group-Ganglioglioma-with-pedLGG" data-toc-modified-id="Group-Ganglioglioma-with-pedLGG-3.3">Group Ganglioglioma with pedLGG</a></span></li><li><span><a href="#Group-DNET-with-pedLGG" data-toc-modified-id="Group-DNET-with-pedLGG-3.4">Group DNET with pedLGG</a></span></li><li><span><a href="#Group-SEGA-(from-Others-group)-with-pedLGG" data-toc-modified-id="Group-SEGA-(from-Others-group)-with-pedLGG-3.5">Group SEGA (from Others group) with pedLGG</a></span></li><li><span><a href="#Group-PA-(ICGC-data)-with-pedLGG" data-toc-modified-id="Group-PA-(ICGC-data)-with-pedLGG-3.6">Group PA (ICGC data) with pedLGG</a></span></li><li><span><a href="#Remove-met-NBL-cases-from-CBTTC-dataset:-7316-224-and-7316-3311" data-toc-modified-id="Remove-met-NBL-cases-from-CBTTC-dataset:-7316-224-and-7316-3311-3.7">Remove met NBL cases from CBTTC dataset: 7316-224 and 7316-3311</a></span></li><li><span><a href="#Bucket-DIPGs-into-pedHGG" data-toc-modified-id="Bucket-DIPGs-into-pedHGG-3.8">Bucket DIPGs into pedHGG</a></span></li><li><span><a href="#Reannotate-Choroid-plexus-as-CP" data-toc-modified-id="Reannotate-Choroid-plexus-as-CP-3.9">Reannotate Choroid plexus as CP</a></span></li><li><span><a href="#Reannotate-Ewings-as-ES" data-toc-modified-id="Reannotate-Ewings-as-ES-3.10">Reannotate Ewings as ES</a></span></li><li><span><a href="#Reannotate-two-embryonal-tumours-as-ETMR" data-toc-modified-id="Reannotate-two-embryonal-tumours-as-ETMR-3.11">Reannotate two embryonal tumours as ETMR</a></span></li><li><span><a href="#Cleanup-Others-group-by-path-reports-and-remove-remaining-samples" data-toc-modified-id="Cleanup-Others-group-by-path-reports-and-remove-remaining-samples-3.12">Cleanup Others group by path reports and remove remaining samples</a></span></li><li><span><a href="#Remove-PNET-group" data-toc-modified-id="Remove-PNET-group-3.13">Remove PNET group</a></span></li><li><span><a href="#Check-if-all-ATRT-cases-have-SMARCB1-mutation" data-toc-modified-id="Check-if-all-ATRT-cases-have-SMARCB1-mutation-3.14">Check if all ATRT cases have SMARCB1 mutation</a></span></li><li><span><a href="#Modify-cohort-abbreviations" data-toc-modified-id="Modify-cohort-abbreviations-3.15">Modify cohort abbreviations</a></span></li></ul></li></ul></div>

# Dependencies

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

load(file = paste0(datapath,"ESTIMATE/estimate_manifest_primary.RData"))

estimate_manifest_primary_clean <- estimate_manifest_primary

## Cleanup according to CRC review

Checked with CRC, see excel file CBTTC Dataset_pathrepchecked_CRCreviewed in Manifests

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-740"] <- "EP WHO Grade III"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-740"] <- "EP"

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-952"] <- "Composite DNET and GGM"

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-715"] <- "GBM"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-715"] <- "pedHGG"

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-1855"] <- "Burkitts Lymphoma"

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-2153"] <- "Hemangiopericytoma"

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-3817"] <- "GBM"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-3817"] <- "pedHGG"

## Modify disease types based on my path reviews

Bucket DNET, GG, PA, PXA and otherLGG into pedLGG

See excel sheet CBTTC Dataset_pathrepchecked_primary

pathchecked <- read.csv(file = paste0(manifestpath, "CBTTC Dataset_pathrepchecked_primary.csv"),
                        header = T, na.strings = "", stringsAsFactors = F)

pathchecked$Cohort[pathchecked$CBTTC.Event.ID == "7316-740"] <- "EP"

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-373"] <- "Ganglioglioma Grade III"
estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-2980"] <- "Malignant glioma with features of PXA"
estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-3922"] <- "Anaplastic PA"

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-373" |
                                       estimate_manifest_primary_clean$sample_id == "7316-2980" |
                                       estimate_manifest_primary_clean$sample_id == "7316-3922"] <- "pedHGG"

pxas <- pathchecked$CBTTC.Event.ID[grepl("PXA",pathchecked$Final.Dx.on.path.report)]
pxas <- pxas[pxas != "7316-2980"]

pas <- pathchecked$CBTTC.Event.ID[grepl("PA",pathchecked$Final.Dx.on.path.report)|
                                 grepl("ilomyxoid",pathchecked$Final.Dx.on.path.report)]
pas <- pas[pas != "7316-3922"]

otherlggs <- pathchecked$CBTTC.Event.ID[pathchecked$Cohort == "LGG" &
                                       !pathchecked$CBTTC.Event.ID %in% pxas &
                                       !pathchecked$CBTTC.Event.ID %in% pas]
otherlggs <- otherlggs[otherlggs != "7316-3922"]

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id %in% pxas] <- "PXA"
estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id %in% pas] <- "PA"

If path report was available for otherlggs, add disease type, if not, keep as is: Low grade glioma/astrocytoma WHO grade I/II

myotherlggdx <- pathchecked[pathchecked$CBTTC.Event.ID %in% otherlggs,]
myotherlggdx <- myotherlggdx[myotherlggdx$Final.Dx.on.path.report != "na",]

for(i in 1:nrow(myotherlggdx)){
estimate_manifest_primary_clean$disease_type[match(myotherlggdx$CBTTC.Event.ID[i], 
                                                   estimate_manifest_primary_clean$sample_id)] <-  myotherlggdx$Final.Dx.on.path.report[i]    
}

## Group Ganglioglioma with pedLGG

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Ganglioglioma"] <- "pedLGG"

## Group DNET with pedLGG

estimate_manifest_primary_clean$cohort[grepl("DNET",estimate_manifest_primary_clean$disease_type)] <- "pedLGG"

## Group SEGA (from Others group) with pedLGG 

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Glial-neuronal tumor NOS"] <- "pedLGG"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Subependymal Giant Cell Astrocytoma (SEGA)"] <- "pedLGG"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Papillary Glioneuronal"] <- "pedLGG"

## Group PA (ICGC data) with pedLGG 

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$cohort == "PA"] <- "pedLGG"

## Remove met NBL cases from CBTTC dataset: 7316-224 and 7316-3311

estimate_manifest_primary_clean <- estimate_manifest_primary_clean[estimate_manifest_primary_clean$sample_id != "7316-3311" &
                                                                   estimate_manifest_primary_clean$sample_id != "7316-224",]

## Bucket DIPGs into pedHGG 

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Brainstem glioma- Diffuse intrinsic pontine glioma"] <- "pedHGG"

## Reannotate Choroid plexus as CP 

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Choroid plexus carcinoma"] <- "CP"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Choroid plexus papilloma"] <- "CP"

## Reannotate Ewings as ES 

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Ewings Sarcoma"] <- "ES"

## Reannotate two embryonal tumours as ETMR

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$cohort == "ET"] <- "ETMR"

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Medulloepithelioma"] <- "ETMR"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$disease_type == "Embryonal Tumor with Multilayer Rosettes, ROS (WHO Grade IV)"] <- "ETMR"

## Cleanup Others group by path reports and remove remaining samples

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-1845"] <- "astrocytoma fibrillary type with intrinsic vascular malformation with necrosis"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-1845"] <- "pedLGG"

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-133"] <- "pedLGG"

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-157"] <- "Atypical DNET/GG/Cortical dysplasia"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-157"] <- "pedLGG"

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-2495"] <- "pedLGG"

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-2513"] <- "pedLGG"

estimate_manifest_primary_clean$disease_type[estimate_manifest_primary_clean$sample_id == "7316-71"] <- "Gliomatosis cerebri with extensive anaplasia"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$sample_id == "7316-71"] <- "pedHGG"

estimate_manifest_primary_clean <- estimate_manifest_primary_clean[estimate_manifest_primary_clean$cohort != "Other",]

## Remove PNET group

estimate_manifest_primary_clean <- estimate_manifest_primary_clean[estimate_manifest_primary_clean$cohort != "PNET",]

## Check if all ATRT cases have SMARCB1 mutation

ATRTsample_ids <- estimate_manifest_primary_clean$sample_id[estimate_manifest_primary_clean$cohort == "ATRT" &
                                                           estimate_manifest_primary_clean$group == "CBTTC"]

I checked all path reports, see excel sheet CBTTC Dataset_Arash Nabbi 12.04.18_pathrepchecked. The majority of ATRT samples either had loss of INI1 expression assessed by IHC or had genetic test (eg SNParray) confirming the deletion of SMARCB1 locus. Exceptions are the following:


7316-479

7316-1073: No path report

7316-2090

7316-3937

I will remove them...

estimate_manifest_primary_clean <- estimate_manifest_primary_clean[estimate_manifest_primary_clean$sample_id != "7316-479" &
                                                                   estimate_manifest_primary_clean$sample_id != "7316-1073" &
                                                                   estimate_manifest_primary_clean$sample_id != "7316-2090" &
                                                                   estimate_manifest_primary_clean$sample_id != "7316-3937",]

## Modify cohort abbreviations

table(estimate_manifest_primary_clean$cohort)

estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$cohort == "MN"] <- "MNG"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$cohort == "SCHN"] <- "SCHW"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$cohort == "EP"] <- "EPN"
estimate_manifest_primary_clean$cohort[estimate_manifest_primary_clean$cohort == "ES"] <- "EWS"

table(estimate_manifest_primary_clean$cohort)

Remove EWS and TT/GN. They are less than 10% of the entire cohort.

estimate_manifest_primary_clean <- 
estimate_manifest_primary_clean[!estimate_manifest_primary_clean$cohort %in% c("EWS", "TT/GN"),]

table(estimate_manifest_primary_clean$group)

dim(estimate_manifest_primary_clean)

table(estimate_manifest_primary_clean$group)

Replace CBTTC with CBTN and DKFZ with ICGC

estimate_manifest_primary_clean$group[estimate_manifest_primary_clean$group == "DKFZ"] <- "ICGC"
estimate_manifest_primary_clean$group[estimate_manifest_primary_clean$group == "CBTTC"] <- "CBTN"

save(estimate_manifest_primary_clean, 
    file = paste0(datapath,"ESTIMATE/estimate_manifest_primary_clean.RData"))
