==> notebooks/06_exp_matrix_4clustering.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Functions" data-toc-modified-id="Functions-2">Functions</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-3">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-4">Main</a></span><ul class="toc-item"><li><span><a href="#CBTN" data-toc-modified-id="CBTN-4.1">CBTN</a></span></li><li><span><a href="#ICGC" data-toc-modified-id="ICGC-4.2">ICGC</a></span></li><li><span><a href="#TARGET" data-toc-modified-id="TARGET-4.3">TARGET</a></span></li><li><span><a href="#TCGA" data-toc-modified-id="TCGA-4.4">TCGA</a></span></li><li><span><a href="#Merge-all" data-toc-modified-id="Merge-all-4.5">Merge all</a></span></li></ul></li></ul></div>

# Dependencies

library(tidyr)
library(biomaRt)

# Functions

matchexp_matrixfx <- function(exp_matrix, estimate_manifest, group, sample_type, matchingcol){
    
    message("dimensions of exp_matrix: ", deparse(substitute(exp_matrix)))
    print(dim(exp_matrix)) 
    subset_estimate_manifest <- estimate_manifest[estimate_manifest$group == group,]
    subset_estimate_manifest_sampletype <- subset_estimate_manifest[grepl(sample_type, subset_estimate_manifest$sample_type),]
    message("dimensions of estimate_manifest_df for: ", group, " sample_type:", sample_type)
    print(dim(subset_estimate_manifest_sampletype)) 
    exp_matrix_matchingset <- exp_matrix[,colnames(exp_matrix) %in% subset_estimate_manifest_sampletype[[matchingcol]]] 
    message("dimensions of exp_matrix_matchingset for: ", deparse(substitute(exp_matrix)), " with matching column: ", matchingcol)
    print(dim(exp_matrix_matchingset))
    exp_matrix_matchingset <- cbind(exp_matrix[,1:2], exp_matrix_matchingset)
    
    return(exp_matrix_matchingset)
}

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

Subset to primary samples

## CBTN

This file is post cleanup, post estimate runs, post primary subset, pre clustering. for expression_clustering

CBTTCtpm <- read.table(paste0(datapath, "exp_mat/CBTTC_tpm_matrix_dedup.txt"), header = T, sep = "\t", stringsAsFactors = F, check.names = F)

CBTTCtpm[1:10,1:10]

load(file = paste0(datapath, "/ESTIMATE/estimate_manifest_primary_clean.RData"))

table(estimate_manifest_primary_clean$group)

CBTTC_tpm_matrix_primary <- matchexp_matrixfx(CBTTCtpm, estimate_manifest_primary_clean,"CBTN", "Initial", "sample_id")

CBTTC_tpm_matrix_primary$ensembl_id <- gsub("[.].*", "", CBTTC_tpm_matrix_primary$ensembl_id)

CBTN_tpm_matrix_primary_dedup <- CBTTC_tpm_matrix_primary[!duplicated(CBTTC_tpm_matrix_primary$ensembl_id),]

dim(CBTTC_tpm_matrix_primary)
dim(CBTN_tpm_matrix_primary_dedup)

write.table(CBTN_tpm_matrix_primary_dedup,
            file = paste0(datapath, "exp_mat/CBTTC_tpm_matrix_primary_dedup.txt"), sep = "\t", quote = T, row.names = F)

## ICGC

ICGC <- read.table(file = paste0(datapath,"exp_mat/ICGC.tpm.matrix.txt"),
                   sep = "\t", header = TRUE, stringsAsFactors = FALSE)

dim(ICGC)

ICGC[1:10,1:10]

colnames(ICGC)[1] <- "ensembl_id"

Some cleanup for colnames

colnames(ICGC) <- gsub(".genes.results", "", colnames(ICGC))
colnames(ICGC) <- gsub("[.]", "-", colnames(ICGC))

dim(ICGC)

head(ICGC)

Remove PAR_Y pseudoautosomal regions

ICGC <- ICGC[!grepl("PAR_Y", ICGC$ensembl_id),]

dim(ICGC)

ICGC$ensembl_id <- gsub("[.].*", "", ICGC$ensembl_id)

ICGC_dedup <- ICGC[ !duplicated(ICGC$ensembl_id),]

dim(ICGC_dedup)

ICGC_exp_matrix_primary <- matchexp_matrixfx(ICGC_dedup, estimate_manifest_primary_clean,"ICGC", "Primary", "sample_id")

save(ICGC_exp_matrix_primary,
     file = paste0(datapath, "exp_mat/DKFZ_tpm_matrix_primary.RData"))

## TARGET

TARGET_tpm_matrix <- read.table(file = paste0(datapath,"exp_mat/NBL_tpm_matrix_hugo.txt"),
                                sep = "\t", header = TRUE, stringsAsFactors = FALSE, check.names = F)

dim(TARGET_tpm_matrix)

TARGET_tpm_matrix_primary <- matchexp_matrixfx(TARGET_tpm_matrix, estimate_manifest_primary_clean,
                                              "TARGET", "Primary", "sample_id")

Remove duplicated genes

TARGET_tpm_matrix_primary$ensembl <- gsub("[.].*", "", TARGET_tpm_matrix_primary$ensembl)

TARGET_tpm_matrix_primary_dedup <- TARGET_tpm_matrix_primary[!duplicated(TARGET_tpm_matrix_primary$ensembl),]

dim(TARGET_tpm_matrix_primary_dedup)

save(TARGET_tpm_matrix_primary_dedup,
         file = paste0(datapath, "exp_mat/TARGET_tpm_matrix_primary_dedup.RData"))

NBL matrix hgnc only

TARGET_tpm_matrix_primary_dedup_hgnc <- TARGET_tpm_matrix_primary[!duplicated(TARGET_tpm_matrix_primary$gene_symbol),]
rownames(TARGET_tpm_matrix_primary_dedup_hgnc) <- TARGET_tpm_matrix_primary_dedup_hgnc$gene_symbol
TARGET_tpm_matrix_primary_dedup_hgnc$gene_symbol <- NULL
TARGET_tpm_matrix_primary_dedup_hgnc$ensembl <- NULL

head(TARGET_tpm_matrix_primary_dedup_hgnc)

write.table(TARGET_tpm_matrix_primary_dedup_hgnc,
                file = paste0(datapath, "exp_mat/tpm_matrix_ped_TARGET_NBL_HGNConly.txt"),
                sep = "\t", quote = F, row.names = T)

## TCGA

TCGA_tpm_matrix <- read.table(file = paste0(datapath,"exp_mat/TCGA_tpm_matrix_hugo.txt"),
                              sep = "\t", header = TRUE, stringsAsFactors = FALSE, check.names = F)

head(TCGA_tpm_matrix)

TCGA_tpm_matrix_primary <- matchexp_matrixfx(TCGA_tpm_matrix, estimate_manifest_primary_clean,"TCGA", "Primary", "sample_id")

Remove duplicated ensembl id (if any)

TCGA_tpm_matrix_primary$ensembl <- gsub("[.].*", "", TCGA_tpm_matrix_primary$ensembl)

TCGA_tpm_matrix_primary_dedup <- TCGA_tpm_matrix_primary[!duplicated(TCGA_tpm_matrix_primary$ensembl),]

dim(TCGA_tpm_matrix_primary_dedup)

save(TCGA_tpm_matrix_primary_dedup,
       file = paste0(datapath, "exp_mat/TCGA_tpm_matrix_primary_dedup.RData"))

## Merge all

colnames(TARGET_tpm_matrix_primary_dedup)[1:2] 
colnames(TCGA_tpm_matrix_primary_dedup)[1:2] 
colnames(ICGC_exp_matrix_primary)[1:2] 
colnames(CBTN_tpm_matrix_primary_dedup)[1:2]

colnames(TARGET_tpm_matrix_primary_dedup)[1]  <- "ensembl_id"
colnames(TCGA_tpm_matrix_primary_dedup)[1]  <- "ensembl_id"

merge with ensembl ids

tpm_matrix_ped <- merge(CBTN_tpm_matrix_primary_dedup, TARGET_tpm_matrix_primary_dedup, by = "ensembl_id")

dim(CBTN_tpm_matrix_primary_dedup)
dim(TARGET_tpm_matrix_primary_dedup)
dim(tpm_matrix_ped)

tpm_matrix_ped <- merge(tpm_matrix_ped, ICGC_exp_matrix_primary, by = "ensembl_id")

dim(ICGC_exp_matrix_primary)
dim(tpm_matrix_ped)

head(tpm_matrix_ped)

dim(tpm_matrix_ped)

rownames(tpm_matrix_ped) <- tpm_matrix_ped$ensembl_id

tpm_matrix_ped$gene_symbol <- NULL
tpm_matrix_ped$ensembl_id <- NULL

head(tpm_matrix_ped)

remove non-coding-RNA and pseudogenes

hg38 <- useMart(biomart="ENSEMBL_MART_ENSEMBL", host="www.ensembl.org", 
                path="/biomart/martservice", dataset="hsapiens_gene_ensembl")

ensembls <- rownames(tpm_matrix_ped)

ensembl_hgnc_type <- getBM(filters="ensembl_gene_id", 
                           attributes=c("hgnc_symbol","ensembl_gene_id", "entrezgene_id", "gene_biotype"), 
                           values= ensembls, mart=hg38)

table(ensembl_hgnc_type$gene_biotype, useNA = "always")

biotypes <- as.data.frame(table(ensembl_hgnc_type$gene_biotype), stringsAsFactors = F)$Var1

biotypes_pseudogenes <- biotypes[grepl("pseudo", biotypes)]
biotypes_RNA <- biotypes[grepl("RNA", biotypes)]

biotypes_RNA

biotypes_pseudogenes

rm_genes <- ensembl_hgnc_type$ensembl_gene_id[ensembl_hgnc_type$gene_biotype %in% c(biotypes_RNA,"TEC", biotypes_pseudogenes)]

dim(tpm_matrix_ped)

tpm_matrix_ped <- tpm_matrix_ped[!rownames(tpm_matrix_ped) %in% rm_genes,]

dim(tpm_matrix_ped)

tpm_matrix_ped_ensembl <- tpm_matrix_ped[, colnames(tpm_matrix_ped) != "hgnc_symbol"]

head(tpm_matrix_ped_ensembl)

save(tpm_matrix_ped_ensembl,
         file = paste0(datapath, "exp_mat/tpm_matrix_ped_primary_dedup_ensembl.RData"))  

rownames(tpm_matrix_ped) <- tpm_matrix_ped$hgnc_symbol
tpm_matrix_ped$hgnc_symbol <- NULL

head(tpm_matrix_ped)

dim(tpm_matrix_ped)

save(tpm_matrix_ped,
     file = paste0(datapath, "exp_mat/tpm_matrix_ped_primary_dedup.RData"))   

write.table(tpm_matrix_ped,
            file = paste0(datapath, "exp_mat/tpm_matrix_ped_primary_dedup.txt"), sep = "\t", quote = F)
