==> notebooks/08_CBTN_Sankey.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-2">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-3">Main</a></span><ul class="toc-item"><li><span><a href="#Remove-cell-lines" data-toc-modified-id="Remove-cell-lines-3.1">Remove cell lines</a></span></li><li><span><a href="#Duplicated-samples" data-toc-modified-id="Duplicated-samples-3.2">Duplicated samples</a></span></li><li><span><a href="#Sample-types" data-toc-modified-id="Sample-types-3.3">Sample types</a></span></li><li><span><a href="#Sufficient-sample-size" data-toc-modified-id="Sufficient-sample-size-3.4">Sufficient sample size</a></span></li><li><span><a href="#Clustering" data-toc-modified-id="Clustering-3.5">Clustering</a></span></li></ul></li><li><span><a href="#Sankey-plot" data-toc-modified-id="Sankey-plot-4">Sankey plot</a></span></li></ul></div>

# Dependencies

library(ggforce)

source("~/OneDrive - UHN/R_src/ggplot2_theme.R")

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

Start with RSEM manifest file 

CBTN_RNA_manifest_genes <- read.csv(paste0(manifestpath, "CBTTC_RSEM-manifest.csv"),header = TRUE, stringsAsFactors = FALSE)

dim(CBTN_RNA_manifest_genes)

CBTN_samples <- cbind.data.frame(CBTN_RNA_manifest_genes$sample_id,
                                 CBTN_RNA_manifest_genes$aliquot_id, stringsAsFactors = F)
colnames(CBTN_samples) <- c("CBTN_sample_id", "CBTN_aliquot_id")

CBTN_samples$CBTN_sample_aliquot <- paste0(CBTN_samples$CBTN_sample_id,"_", CBTN_samples$CBTN_aliquot_id)

CBTN_samples$CBTN <- "CBTN\nsamples\n(n = 996)"

head(CBTN_samples)

## Remove cell lines

CBTN_celllines <- read.csv(paste0(manifestpath,"cell_line_RNAalqids_CBTTC.csv"), header = TRUE, stringsAsFactors = F)

head(CBTN_celllines)

CBTN_celllines$sample_aliquot_id <- paste0(CBTN_celllines$clinical_event, "_", CBTN_celllines$cell_lines_rnaalq_id)

cellines <- CBTN_samples$CBTN_sample_aliquot[ CBTN_samples$CBTN_sample_aliquot %in% CBTN_celllines$sample_aliquot_id]
not_celllines <- CBTN_samples$CBTN_sample_aliquot[!CBTN_samples$CBTN_sample_aliquot %in% CBTN_celllines$sample_aliquot_id]

cellines

CBTN_samples$Cell_lines <- NA
CBTN_samples$Cell_lines[ CBTN_samples$CBTN_sample_aliquot %in% cellines] <- 
paste0("Cell lines\n(n = ", length(cellines), ")")

CBTN_samples$Cell_lines[CBTN_samples$CBTN_sample_aliquot %in% not_celllines] <- 
paste0("Tumours\n(n = ", length(not_celllines), ")")

table(CBTN_samples$Cell_lines)

head(CBTN_samples)

## Duplicated samples

CBTN_RNA_manifest_genes$sample_aliquot_id <- paste0(CBTN_RNA_manifest_genes$sample_id, "_", CBTN_RNA_manifest_genes$aliquot_id)

CBTN_RNA_manifest_genes_linerm <- CBTN_RNA_manifest_genes[! CBTN_RNA_manifest_genes$sample_aliquot_id %in% cellines,]

dim(CBTN_RNA_manifest_genes_linerm)

CBTN_RNA_manifest_genes_linerm_dedup <- CBTN_RNA_manifest_genes_linerm[!duplicated(CBTN_RNA_manifest_genes_linerm$sample_id),]

dedup_ids <- CBTN_RNA_manifest_genes_linerm_dedup$sample_aliquot_id

dim(CBTN_RNA_manifest_genes_linerm_dedup)

dup_ids <- CBTN_RNA_manifest_genes_linerm$sample_aliquot_id[duplicated(CBTN_RNA_manifest_genes_linerm$sample_id)]

dup_ids

CBTN_samples$duplicated_samples <- NA

CBTN_samples$duplicated_samples[CBTN_samples$CBTN_sample_aliquot %in% dup_ids] <- 
paste0("Duplicated (n = ",length(dup_ids), ")") 
CBTN_samples$duplicated_samples[CBTN_samples$CBTN_sample_aliquot %in% dedup_ids] <-
paste0("Unique\n(n = ",length(dedup_ids), ")") 

table(CBTN_samples$duplicated_samples)

## Sample types

load(file = paste0(manifestpath,"Master_manifest.RData"))

cbtn_metadata <- IPD_metadata[ IPD_metadata$group == "CBTTC",]

table(cbtn_metadata$sample_type, useNA = "always")

cbtn_metadata$sample_aliquot_id <- paste0(cbtn_metadata$sample_id, "_", cbtn_metadata$aliquot_id)

primaries <- cbtn_metadata$sample_aliquot_id[cbtn_metadata$sample_type == "Initial CNS Tumor"]
progressives <- cbtn_metadata$sample_aliquot_id[cbtn_metadata$sample_type== "Progressive"]
recurrences <- cbtn_metadata$sample_aliquot_id[cbtn_metadata$sample_type == "Recurrence"]
secondmalig <- cbtn_metadata$sample_aliquot_id[cbtn_metadata$sample_type == "Second Malignancy"]
notavailables <- cbtn_metadata$sample_aliquot_id[cbtn_metadata$sample_type %in% c("","Unavailable")]

length(primaries)

CBTN_samples$tumour_type <- NA

CBTN_samples$tumour_type[ CBTN_samples$CBTN_sample_aliquot %in% primaries] <- 
paste0("Primary\n(n = ", 
       length(primaries), ")")

CBTN_samples$tumour_type[ CBTN_samples$CBTN_sample_aliquot %in% progressives] <- 
paste0("Progressive\n(n = ", 
       length(progressives), ")")

CBTN_samples$tumour_type[ CBTN_samples$CBTN_sample_aliquot %in% recurrences ] <- 
paste0("Recurrence (n = ", 
       length(recurrences), ")")

CBTN_samples$tumour_type[ CBTN_samples$CBTN_sample_aliquot %in% secondmalig ] <- 
paste0("Second malignancy\n(n = ", 
       length(secondmalig), ")")

table(CBTN_samples$tumour_type, useNA = "always")

## Sufficient sample size

load(file = paste0(datapath,"ESTIMATE/estimate_manifest_primary.RData"))

load(file = paste0(datapath,"ESTIMATE/estimate_manifest_primary_clean.RData"))

cbtn_primary <- estimate_manifest_primary[ estimate_manifest_primary$group == "CBTTC",]

cbtn_primary_clean <- estimate_manifest_primary_clean[ estimate_manifest_primary_clean$group == "CBTN",]

dim(cbtn_primary)
dim(cbtn_primary_clean)

cbtn_primary$sample_aliquot_id <- paste0(cbtn_primary$sample_id, "_", cbtn_primary$aliquot_id)
cbtn_primary_clean$sample_aliquot_id <- paste0(cbtn_primary_clean$sample_id, "_", cbtn_primary_clean$aliquot_id)

goodsamplesize <- cbtn_primary_clean$sample_aliquot_id
badsamplesize <- cbtn_primary$sample_aliquot_id[ ! cbtn_primary$sample_aliquot_id %in% cbtn_primary_clean$sample_aliquot_id]

CBTN_samples$sufficient_samplesize <- NA

CBTN_samples$sufficient_samplesize[ CBTN_samples$CBTN_sample_aliquot %in% goodsamplesize ] <- 
paste0("Sufficient\nsample size\n(n = ", length(goodsamplesize), ")")
CBTN_samples$sufficient_samplesize[ CBTN_samples$CBTN_sample_aliquot %in% badsamplesize ] <- 
paste0("Rare/unannotated\n(n = ", length(badsamplesize), ")")

table(CBTN_samples$sufficient_samplesize, useNA = "always")

## Clustering

load(file = paste0(datapath, "/ESTIMATE/estimate_manifest_primary_clean.RData"))

cbtn_preclustering <- estimate_manifest_primary_clean[ estimate_manifest_primary_clean$group == "CBTN",]

load(file = paste0(datapath, "ESTIMATE/estimate_manifest_primary_clean_postclustering.RData"))  

cbtn_postclustering <- estimate_manifest_primary_clean[ estimate_manifest_primary_clean$group == "CBTN",]

cbtn_preclustering$sample_aliquot_id <- paste0(cbtn_preclustering$sample_id, "_",cbtn_preclustering$aliquot_id)

cbtn_postclustering$sample_aliquot_id <- paste0(cbtn_postclustering$sample_id, "_",cbtn_postclustering$aliquot_id)

dim(cbtn_preclustering)

goodexpression <- cbtn_postclustering$sample_aliquot_id
badexpression <- cbtn_preclustering$sample_aliquot_id[!cbtn_preclustering$sample_aliquot_id %in% cbtn_postclustering$sample_aliquot_id]

length(goodexpression)
length(badexpression)

CBTN_samples$tumorexpression <- NA

CBTN_samples$tumorexpression[ CBTN_samples$CBTN_sample_aliquot %in% goodexpression ] <- paste0("Matched with\npathology\n(n = ",length(goodexpression), ")")

CBTN_samples$tumorexpression[ CBTN_samples$CBTN_sample_aliquot %in% badexpression ] <- 
paste0("Not matched with\npathology\n(n = ",length(badexpression), ")")

table(CBTN_samples$tumorexpression, useNA = "always")

# Sankey plot

CBTN_samples$duplicated_samples[is.na(CBTN_samples$duplicated_samples)] <- "Excluded (n = 22)"
CBTN_samples$tumour_type[is.na(CBTN_samples$tumour_type)] <- "Excluded (n = 27)"
CBTN_samples$sufficient_samplesize[is.na(CBTN_samples$sufficient_samplesize)] <- "Excluded\n(n = 266)"
CBTN_samples$tumorexpression[is.na(CBTN_samples$tumorexpression)] <- "Excluded\n(n = 364)"

freqtab <- as.data.frame(table(CBTN_samples$CBTN, CBTN_samples$Cell_lines, CBTN_samples$duplicated_samples,
                              CBTN_samples$tumour_type, CBTN_samples$sufficient_samplesize,
                              CBTN_samples$tumorexpression), stringsAsFactors = F)

head(freqtab)

freqtab <- freqtab[ order(freqtab$Var2, freqtab$Var3, freqtab$Var4, freqtab$Var5, freqtab$Var6),]

freqtab$Var7 <- NA

freqtab$Var7[ freqtab$Freq == 581] <- "Included\nin the study\n(n = 581)"
freqtab$Var7[ freqtab$Freq != 581] <- "Not included\nin the study\n(n = 415)"

freqtab_ggforce <- gather_set_data(freqtab[freqtab$Freq > 0,], c(1:6,8))

table(freqtab_ggforce$y)

freqtab_ggforce$y <- factor(freqtab_ggforce$y,
levels = c('CBTN\nsamples\n(n = 996)',
           'Tumours\n(n = 974)','Cell lines\n(n = 22)',
           'Unique\n(n = 969)',
           'Duplicated (n = 5)',
           'Excluded (n = 22)',
           'Primary\n(n = 730)',
           'Progressive\n(n = 153)','Recurrence (n = 70)',
           'Second malignancy\n(n = 16)', 'Excluded (n = 27)',
           'Sufficient\nsample size\n(n = 632)', 'Rare/unannotated\n(n = 98)',
           'Excluded\n(n = 266)',
           'Matched with\npathology\n(n = 581)',
           'Not matched with\npathology\n(n = 51)',
           'Excluded\n(n = 364)',
           'Included\nin the study\n(n = 581)',
           'Not included\nin the study\n(n = 415)'
           ))

freqtab_ggforce$includedsamples[ freqtab_ggforce$Freq == 581] <- "Yes"
freqtab_ggforce$includedsamples[ freqtab_ggforce$Freq != 581] <- "No"

p_plot <- ggplot(freqtab_ggforce, aes(x, id = id, split = y, value = Freq)) +
geom_parallel_sets(aes(fill = includedsamples), alpha = 0.5) +
geom_parallel_sets_axes(axis.width = 0.3, fill = "dark gray") +
geom_parallel_sets_labels(colour = 'black', size = 8, angle = 0) + 
myaxis + myplot + 
theme(plot.margin = unit(c(1,10,1,1),"cm"),
      legend.position = "none", 
      axis.title.x = element_blank(), 
      axis.text.x = element_blank(),
      axis.line.x = element_blank(),
      axis.ticks.x = element_blank()) +
scale_y_continuous(expand = c(0.05, 0.5)) + 
scale_x_discrete(expand = c(0.1, 0.1)) +
scale_fill_manual(values = c("Yes" = "#d7191c", "No" = "#2c7bb6"))

pdf(paste0(plotpath, "CBTN_alluvial.pdf"),
   width = 30, height = 10)
p_plot
   
dev.off()
