==> notebooks/10_TMB_dataframe.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-2">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-3">Main</a></span><ul class="toc-item"><li><span><a href="#CBTN-MAF-summary" data-toc-modified-id="CBTN-MAF-summary-3.1">CBTN MAF summary</a></span></li><li><span><a href="#GDC-MAF-summary" data-toc-modified-id="GDC-MAF-summary-3.2">GDC MAF summary</a></span></li><li><span><a href="#DKFZ" data-toc-modified-id="DKFZ-3.3">DKFZ</a></span></li><li><span><a href="#Calculate-Tumour-Mutation-Burden-(TMB)" data-toc-modified-id="Calculate-Tumour-Mutation-Burden-(TMB)-3.4">Calculate Tumour Mutation Burden (TMB)</a></span></li></ul></li></ul></div>

Calculation of TMB from overlapped calls of strelka2 and mutect2  and VAF > 0.03 in CBTTC dataset and mutect2 calls with VAF > 0.03 in GDC dataset

# Dependencies

library(plyr)

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

load(paste0(datapath, "ESTIMATE/estimate_manifest_primary_clean_final.RData"))

table(estimate_manifest_primary_clean$recurrence)

## CBTN MAF summary 

post-recoding post-overlap post-VAF filter

CBTN_mut <- read.table(file = paste0(datapath,"snv/CBTTC_mafsummary_overlapped_3VAF.txt"),sep = "\t", 
                        check.names = F,header = TRUE, stringsAsFactors = FALSE)

head(CBTN_mut[order(CBTN_mut$total, decreasing = T),])

dim(CBTN_mut)

CBTN <- estimate_manifest_primary_clean[estimate_manifest_primary_clean$group == "CBTN",]

dim(CBTN)

length(CBTN_mut$sample_id[CBTN_mut$sample_id %in% CBTN$sample_id])

CBTN_mut_subset <- CBTN_mut[,c("sample_id", "Frame_Shift_Del", "Frame_Shift_Ins","In_Frame_Del", "In_Frame_Ins",
                                 "Missense_Mutation", "Nonsense_Mutation", "Nonstop_Mutation", "total")]

head(CBTN_mut_subset)

mafsum_CBTN_estimate_manifest <- merge(CBTN, CBTN_mut_subset, by = "sample_id")

dim(mafsum_CBTN_estimate_manifest)

head(mafsum_CBTN_estimate_manifest)

## GDC MAF summary 

post-recoding post-deduping post-VAF filter

GDC_mut <- read.table(file = paste0(datapath, "snv/GDC_mafsummary_3VAF_dedup.txt"),
                      sep = "\t", header = TRUE, stringsAsFactors = FALSE)

head(GDC_mut)

dim(GDC_mut)

GDC <- estimate_manifest_primary_clean[estimate_manifest_primary_clean$group == "TCGA" |
                                       estimate_manifest_primary_clean$group == "TARGET",]

dim(GDC)

colnames(GDC_mut)[colnames(GDC_mut) == "Tumor_Sample_Barcode"] <- "sample_id"

GDC_mut_mutects_estimate_manifest <- merge(GDC, GDC_mut[,colnames(GDC_mut) != "group"],
                                                  by = "sample_id")

head(GDC_mut_mutects_estimate_manifest)

table(GDC_mut_mutects_estimate_manifest$recurrence)

dim(GDC)
dim(GDC_mut)
dim(GDC_mut_mutects_estimate_manifest)

GDC_CBTN_mut_manifest_estimate <- rbind(mafsum_CBTN_estimate_manifest, GDC_mut_mutects_estimate_manifest)

## DKFZ

somatic_DKFZ <- read.table(file = paste0(datapath,"snv/somatic_SNVs_INDELs_inCommonRegions.for_TMB.txt"),
                        sep = "\t", header = FALSE, stringsAsFactors = FALSE)

colnames(somatic_DKFZ) <- c("sample_id", "total")
head(somatic_DKFZ)

DKFZ <- estimate_manifest_primary_clean[estimate_manifest_primary_clean$group == "ICGC",]

DKFZ_mut_estimate_manifest <- merge(DKFZ,somatic_DKFZ,by = "sample_id")

dim(DKFZ_mut_estimate_manifest)

total_mut_manifest_estimate <- rbind.fill(GDC_CBTN_mut_manifest_estimate,DKFZ_mut_estimate_manifest)

dim(GDC_CBTN_mut_manifest_estimate)
dim(DKFZ_mut_estimate_manifest)
dim(total_mut_manifest_estimate)

table(total_mut_manifest_estimate$group)

table(total_mut_manifest_estimate$sample_type)

## Calculate Tumour Mutation Burden (TMB)

size of bedfile used to subset variant calls (hglft_agil_VCR_Sure_S029_SeqCapV2V3_S04_pads_cexons_sorted_merge.be): 30028393

total_mut_manifest_estimate$mutpermb <- total_mut_manifest_estimate$total/30.028393

total_mut_manifest_estimate$snvpermb <- (total_mut_manifest_estimate$Missense_Mutation +
                                         total_mut_manifest_estimate$Nonsense_Mutation + 
                                         total_mut_manifest_estimate$Nonstop_Mutation)/30.028393

total_mut_manifest_estimate$indelpermb <- (total_mut_manifest_estimate$Frame_Shift_Del +
                                           total_mut_manifest_estimate$Frame_Shift_Ins + 
                                           total_mut_manifest_estimate$In_Frame_Del +
                                           total_mut_manifest_estimate$In_Frame_Ins)/30.028393

colnames(total_mut_manifest_estimate)

write.csv(total_mut_manifest_estimate[,c("sample_id", 'Frame_Shift_Del','Frame_Shift_Ins','In_Frame_Del',
                                         'In_Frame_Ins','Missense_Mutation','Nonsense_Mutation',
                                         'Nonstop_Mutation','total','mutpermb','snvpermb','indelpermb')],
         file = paste0(datapath, "forGangesh/tmb.csv"), row.names = F)

save(total_mut_manifest_estimate,
     file = paste0(datapath, "snv/total_mut_manifest_estimate.RData"))
