==> notebooks/11_tumour_subgroups.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-2">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-3">Main</a></span></li><li><span><a href="#Add-tumour-subtypes" data-toc-modified-id="Add-tumour-subtypes-4">Add tumour subtypes</a></span><ul class="toc-item"><li><span><a href="#ATRT-and-MB" data-toc-modified-id="ATRT-and-MB-4.1">ATRT and MB</a></span></li><li><span><a href="#GBM-subgroups" data-toc-modified-id="GBM-subgroups-4.2">GBM subgroups</a></span><ul class="toc-item"><li><span><a href="#CBTN" data-toc-modified-id="CBTN-4.2.1">CBTN</a></span></li><li><span><a href="#ICGC" data-toc-modified-id="ICGC-4.2.2">ICGC</a></span></li></ul></li><li><span><a href="#EPN-subgroups" data-toc-modified-id="EPN-subgroups-4.3">EPN subgroups</a></span></li></ul></li><li><span><a href="#NBL-subgroups" data-toc-modified-id="NBL-subgroups-5">NBL subgroups</a></span></li><li><span><a href="#LGG-subgroups" data-toc-modified-id="LGG-subgroups-6">LGG subgroups</a></span></li></ul></div>

# Dependencies

library(maftools)

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

load(paste0(datapath, "ESTIMATE/estimate_manifest_primary_clean_final.RData"))

table(estimate_manifest_primary_clean$group)

metadata <- estimate_manifest_primary_clean

# Add tumour subtypes

metadata$tumour_subtype <- NA

## ATRT and MB 

from ICGC or gene expression clustering

metadata$tumour_subtype[ metadata$cohort == "MB"] <- metadata$disease_type[ metadata$cohort == "MB"]

table(metadata$tumour_subtype)

# some cleanup
metadata$tumour_subtype[metadata$tumour_subtype == "MB_Group3"] <- "MB_GR34"
metadata$tumour_subtype[metadata$tumour_subtype == "MB_Group4"] <- "MB_GR34"

table(metadata$tumour_subtype)

metadata$tumour_subtype[ metadata$cohort == "ATRT"] <- metadata$disease_type[ metadata$cohort == "ATRT"]

## GBM subgroups

### CBTN

The following is to add variantfilenames to oncokb annotated combined_maf using CBTTC_summary_mutations_MuTect2.txt (generated on cavatica)

CBTN_mafsummary <- read.table(file = paste0(datapath,"snv/CBTTC_mafsummary_overlapped_3VAF.txt"),
                              sep = "\t", check.names = F, header = TRUE,  stringsAsFactors = FALSE)

head(CBTN_mafsummary)

dim(CBTN_mafsummary)

mymaf_CBTN <- read.maf(maf =  paste0(datapath,"snv/CBTTC_maf_overlapped_3VAF.maf"),
                       useAll = TRUE, verbose = FALSE,
                       vc_nonSyn = c("Frame_Shift_Del", "Frame_Shift_Ins", "Nonsense_Mutation","Nonstop_Mutation", 
                                     "In_Frame_Del","In_Frame_Ins", "Missense_Mutation"))

mymaf_CBTN@summary

CBTN_data_maf <- mymaf_CBTN@data

CBTN_data_maf_sampleid <- merge(CBTN_mafsummary[,c("sample_id", "Tumor_Sample_Barcode")],
                                CBTN_data_maf,by = "Tumor_Sample_Barcode")

length(unique(CBTN_data_maf_sampleid$Tumor_Sample_Barcode))

Subset  snv dfs to include only samples in estimate df

CBTN_data_maf_primary <- CBTN_data_maf_sampleid[CBTN_data_maf_sampleid$sample_id %in% metadata$sample_id,]

length(unique(CBTN_data_maf_primary$Tumor_Sample_Barcode))

head(CBTN_data_maf_primary)

cbtn_h3f3a <- CBTN_data_maf_primary[which(CBTN_data_maf_primary$Hugo_Symbol == "H3F3A"),]
cbtn_idh1 <- CBTN_data_maf_primary[which(CBTN_data_maf_primary$Hugo_Symbol == "IDH1"),]

cbtn_h3f3a_k27 <- cbtn_h3f3a[which(grepl("K28M", cbtn_h3f3a$HGVSp_Short)),]
cbtn_h3f3a_g34 <- cbtn_h3f3a[which(grepl("G35", cbtn_h3f3a$HGVSp_Short)),]

k27_samples <- cbtn_h3f3a_k27$sample_id
g34_samples <- cbtn_h3f3a_g34$sample_id
idh1_samples <- cbtn_idh1$sample_id

metadata$tumour_subtype[metadata$sample_id %in% idh1_samples &
                        metadata$cohort == "pedHGG" & metadata$group == "CBTN"] <- "HGG_IDH1"
metadata$tumour_subtype[metadata$sample_id %in% g34_samples &
                        metadata$cohort == "pedHGG"& metadata$group == "CBTN"] <- "HGG_G34"
metadata$tumour_subtype[metadata$sample_id %in% k27_samples &
                        metadata$cohort == "pedHGG"& metadata$group == "CBTN"] <- "HGG_K27M"
metadata$tumour_subtype[is.na(metadata$tumour_subtype) &
                        metadata$cohort == "pedHGG"& metadata$group == "CBTN"] <- "HGG_WT"

### ICGC

Get DKFZ tumour subtypes from clinical metadata

metadata$tumour_subtype[metadata$cohort == "pedHGG" & metadata$group == "ICGC"] <- 
metadata$disease_type[metadata$cohort == "pedHGG" & metadata$group == "ICGC"]

metadata$disease_type[metadata$cohort == "pedHGG" & metadata$group == "ICGC"]

metadata$tumour_subtype[grepl("HGGother_IDH", metadata$tumour_subtype)] <- "HGG_IDH1"
metadata$tumour_subtype[grepl("HGG_K27M", metadata$tumour_subtype)] <- "HGG_K27M"
metadata$tumour_subtype[grepl("HGGother_G34R", metadata$tumour_subtype)] <- "HGG_G34"
metadata$tumour_subtype[grepl("HGGother_WT", metadata$tumour_subtype)] <- "HGG_WT"

##  EPN subgroups

EPNs <- metadata[metadata$cohort == "EPN",]

table(EPNs$primary_site)

epn_fossa <- EPNs$sample_id[grepl("Posterior Fossa", EPNs$primary_site)]
epn_spine <- EPNs$sample_id[grepl("Spin", EPNs$primary_site)]
epn_supratent <- EPNs$sample_id[grepl("Lobe", EPNs$primary_site) | grepl("Ventricle", EPNs$primary_site)]

metadata$tumour_subtype[metadata$sample_id %in% epn_fossa & metadata$cohort == "EPN"] <- "EPN_Fossa"
metadata$tumour_subtype[metadata$sample_id %in% epn_spine & metadata$cohort == "EPN"] <- "EPN_Spine"
metadata$tumour_subtype[metadata$sample_id %in% epn_supratent & metadata$cohort == "EPN"] <- "EPN_Supra"
metadata$tumour_subtype[is.na(metadata$tumour_subtype) & metadata$cohort == "EPN"] <- "EPN_ND"

# NBL subgroups

From TARGET portal

NBLs <- metadata[metadata$cohort == "NBL",]

NBL_manifest <- read.table(paste0(manifestpath, "TARGET_NBL_Discovery_ClinicalData_20170525.txt"),
                          sep = "\t", na.strings = "", header = TRUE)

colnames(NBL_manifest)[colnames(NBL_manifest) == "TARGET.USI"] <- "case_id"

NBL_manifest_subset <- NBL_manifest[NBL_manifest$case_id %in% NBLs$case_id,]

dim(NBL_manifest_subset)

NBLs_MYCstatus <- merge(NBLs, NBL_manifest_subset[, c("case_id", "MYCN.status")], by = "case_id")

dim(NBLs_MYCstatus)

table(NBLs_MYCstatus$MYCN.status)

head(NBLs_MYCstatus)

nbl_myca <- NBLs_MYCstatus$case_id[NBLs_MYCstatus$MYCN.status == "Amplified"]
nbl_mycna <- NBLs_MYCstatus$case_id[NBLs_MYCstatus$MYCN.status == "Not Amplified"]

metadata$tumour_subtype[metadata$case_id %in% nbl_myca & metadata$cohort == "NBL"] <- "NBL_MYCA"
metadata$tumour_subtype[metadata$case_id %in% nbl_mycna & metadata$cohort == "NBL"] <- "NBL_MYCNA"
metadata$tumour_subtype[is.na(metadata$tumour_subtype)& metadata$cohort == "NBL"] <- "NBL_ND"

table(metadata$tumour_subtype[which(metadata$cohort == "NBL")])

# LGG subgroups

Add BRAF status

load(file = paste0(datapath, "fusion/arriba_df.RData"))

remove low confidence

arriba_df_conf <- arriba_df[arriba_df$confidence != "low",]

braf_fusion_samples <- arriba_df_conf$sample_id[grepl("BRAF", arriba_df_conf$gene2) & grepl("KIAA1549", arriba_df_conf$gene1)]

cbtn_braf <- CBTN_data_maf_primary[which(CBTN_data_maf_primary$Hugo_Symbol == "BRAF"),]
cbtn_braf_v600e_samples <- cbtn_braf$sample_id[which(grepl("V600E", cbtn_braf$HGVSp_Short))]

ICGC_maf <- read.table(file = paste0(datapath, "snv/dkfz_mafKB.maf"),
                        sep = "\t", check.names = F, header = TRUE, comment.char = "",stringsAsFactors = FALSE)

ICGC_maf_nsyn <- ICGC_maf[!ICGC_maf$variant_classification %in% c("splicing", "unknown", "synonymous SNV"),]

colnames(ICGC_maf_nsyn)[colnames(ICGC_maf_nsyn) == "Tumor_Sample_Barcode"] <- "sample_id"

head(ICGC_maf_nsyn)

dkfz_braf <- ICGC_maf_nsyn[which(ICGC_maf_nsyn$Hugo_Symbol == "BRAF"),]
dkfz_braf_v600e_samples <- dkfz_braf$sample_id[which(grepl("V600E", dkfz_braf$HGVSp_Short))]

braf_v600e_samples <- c(dkfz_braf_v600e_samples, cbtn_braf_v600e_samples)

braf_fusion_samples

metadata$tumour_subtype[metadata$sample_id %in% braf_fusion_samples & metadata$cohort == "pedLGG"] <- "BRAF_KIAA1549"
metadata$tumour_subtype[metadata$sample_id %in% braf_v600e_samples & metadata$cohort == "pedLGG"] <- "BRAF_p.V600E"
metadata$tumour_subtype[is.na(metadata$tumour_subtype)& metadata$cohort == "pedLGG"] <- "BRAF_WT"

table(metadata$tumour_subtype)

estimate_manifest_primary_clean <- metadata

save(estimate_manifest_primary_clean,
         file = paste0(datapath, "ESTIMATE/estimate_manifest_primary_clean_final.RData"))
