==> notebooks/12_CRI_clusters.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-2">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-3">Main</a></span></li><li><span><a href="#Add-CRI-clusters" data-toc-modified-id="Add-CRI-clusters-4">Add CRI clusters</a></span></li></ul></div>

# Dependencies

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

load(paste0(datapath,"ESTIMATE/estimate_manifest_primary_clean_final.RData"))

load(paste0(datapath,"ESTIMATE/PDX_estimate.RData"))

table(estimate_manifest_primary_clean$cohort)

table(estimate_manifest_primary_clean$group)

# Add CRI clusters

cri <- read.csv(paste0(datapath, "TME_clusters/immune-calls_CRItool_IPD2.csv"),
               header = T , stringsAsFactors = F, check.names = F)

cri$SampleIDs <- gsub("X", "", cri$SampleIDs)
cri$SampleIDs <- gsub("[.]", "-", cri$SampleIDs)

dim(cri)

colnames(cri)[1] <- "sample_id"
colnames(cri)[2] <- "CRI_cluster"

tail(cri)

estimate_manifest_primary_clean$CRI_cluster <- NA

estimate_manifest_primary_clean$CRI_cluster <- cri$CRI_cluster[ match(estimate_manifest_primary_clean$sample_id,cri$sample_id )]

table(estimate_manifest_primary_clean$group)

save(estimate_manifest_primary_clean, file = paste0(datapath,"ESTIMATE/estimate_manifest_primary_clean_final.RData"))
