==> notebooks/16_TIMER2_heatmap.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Functions" data-toc-modified-id="Functions-2">Functions</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-3">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-4">Main</a></span><ul class="toc-item"><li><span><a href="#Annotations-for-cohorts" data-toc-modified-id="Annotations-for-cohorts-4.1">Annotations for cohorts</a></span></li></ul></li></ul></div>

# Dependencies

options(scipen = 999)

# Functions

source("~/OneDrive - UHN/R_src/ggplot2_theme.R")
source("~/OneDrive - UHN/R_src/color_schemes.R")
source("~/OneDrive - UHN/R_src/Heatmap_functions.R")

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

load(paste0(datapath,"ESTIMATE/estimate_manifest_primary_clean_final.RData"))

ped <- estimate_manifest_primary_clean[ estimate_manifest_primary_clean$group != "TCGA",]

table(ped$group)

load pdx

pdxfilelist <- list.files(paste0(datapath, "TIMER2"), full.names = T, pattern = "estimation_matrix_P")

pdxfiles <- lapply(pdxfilelist, 
                    function(f){read.csv(f, header = TRUE, stringsAsFactors = F, check.names = F, row.names = 1)})

pdxes <- as.data.frame(do.call("cbind", pdxfiles))

pdxes_NS <- pdxes[,grep(paste(c("EP", "MB", "HG", "NB", "RT"),collapse="|"),colnames(pdxes), value=TRUE)]
pdxes_NS <- pdxes_NS[,grep(paste(c("PP", "PT", "PR", "PM"),collapse="|"),colnames(pdxes_NS), value=TRUE)]
pdxes_NS <- pdxes_NS[,!colnames(pdxes_NS) %in% c("ITCC.P4_s07_HG0068_PT01_F01_R01_A03","ITCC.P4_s15_NB0537_PT03_F01_R01_03")]

dim(pdxes_NS)

pdx_mb <- read.csv("/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/TIMER2/estimation_matrix_PDX_MB.csv",
         header = TRUE, stringsAsFactors = F, check.names = F, row.names = 1)

head(pdxes_NS)

pdxes_NS <- cbind(pdxes_NS, pdx_mb)

dim(pdxes_NS)

pdx_cohort <- colnames(pdxes_NS)
names(pdx_cohort) <- pdx_cohort

pdx_cohort[ grepl("EP", pdx_cohort)] <- "EPN"
pdx_cohort[ grepl("NB", pdx_cohort)] <- "NBL"
pdx_cohort[ grepl("RT", pdx_cohort)] <- "ATRT"
pdx_cohort[ grepl("HG", pdx_cohort)] <- "pedHGG"
pdx_cohort[ grepl("MB", pdx_cohort)] <- "MB"

pdx_cohort[ grepl("Med", pdx_cohort)] <- "MB"
pdx_cohort[ grepl("Neo", pdx_cohort)] <- "MB"
pdx_cohort[ grepl("XP", pdx_cohort)] <- "MB"

table(pdx_cohort)

pdxes_NS$Tool <- NA

pdxes_NS$Tool[grepl("TIMER", rownames(pdxes_NS))] <- "TIMER"
pdxes_NS$Tool[grepl("CIBERSORT-ABS", rownames(pdxes_NS))] <- "CIBERSORT"
pdxes_NS$Tool[grepl("EPIC", rownames(pdxes_NS))] <- "EPIC"
pdxes_NS$Tool[grepl("XCELL", rownames(pdxes_NS))] <- "XCELL"
pdxes_NS$Tool[grepl("MCPCOUNTER", rownames(pdxes_NS))] <- "MCPCOUNTER"
pdxes_NS$Tool[grepl("QUANTISEQ", rownames(pdxes_NS))] <- "QUANTISEQ"

pdxes_NS <- pdxes_NS[ !is.na(pdxes_NS$Tool),]

pdxes_NS <- pdxes_NS[! grepl("uncharacterized", rownames(pdxes_NS)),]
pdxes_NS <- pdxes_NS[! grepl("fibroblast", rownames(pdxes_NS)),]
pdxes_NS <- pdxes_NS[! grepl("Endothelial", rownames(pdxes_NS)),]
pdxes_NS <- pdxes_NS[! grepl("Hematopoietic", rownames(pdxes_NS)),]
pdxes_NS <- pdxes_NS[! grepl("progenitor", rownames(pdxes_NS)),]
pdxes_NS <- pdxes_NS[! grepl("score", rownames(pdxes_NS)),]

pediatric cohort

ipdfilelist <- list.files(paste0(datapath, "TIMER2"), full.names = T, pattern = "IPD")

ipdfiles <- lapply(ipdfilelist, 
                    function(f){read.csv(f, header = TRUE, stringsAsFactors = F, check.names = F, row.names = 1)})

ipd <- as.data.frame(do.call("cbind", ipdfiles))

ipd$Tool <- NA

ipd$Tool[grepl("TIMER", rownames(ipd))] <- "TIMER"
ipd$Tool[grepl("CIBERSORT-ABS", rownames(ipd))] <- "CIBERSORT"
ipd$Tool[grepl("EPIC", rownames(ipd))] <- "EPIC"
ipd$Tool[grepl("XCELL", rownames(ipd))] <- "XCELL"
ipd$Tool[grepl("MCPCOUNTER", rownames(ipd))] <- "MCPCOUNTER"
ipd$Tool[grepl("QUANTISEQ", rownames(ipd))] <- "QUANTISEQ"

ipd <- ipd[ !is.na(ipd$Tool),]

rownames(ipd)

write.csv(ipd, paste0(datapath, "forGangesh/TIMER2.csv"))

ipd <- ipd[! grepl("uncharacterized", rownames(ipd)),]
ipd <- ipd[! grepl("fibroblast", rownames(ipd)),]
ipd <- ipd[! grepl("Endothelial", rownames(ipd)),]
ipd <- ipd[! grepl("Hematopoietic", rownames(ipd)),]
ipd <- ipd[! grepl("progenitor", rownames(ipd)),]
ipd <- ipd[! grepl("score", rownames(ipd)),]

head(ipd)

## Annotations for cohorts

ped <- ped[ ped$sample_id != "H049-SPP6",]

Sample above was not available in hg38

mycohort <- ped$cohort
names(mycohort) <- ped$sample_id

length(ped$cohort)

cohort_ha <- HeatmapAnnotation(cohort = mycohort, show_annotation_name = FALSE, col = list(cohort = cohort_col),
                             simple_anno_size = unit(1, "cm"))

pdx_cohort_ha <- HeatmapAnnotation(cohort = pdx_cohort, show_annotation_name = FALSE, col = list(cohort = cohort_col),
                                 simple_anno_size = unit(1, "cm"))

# colors set after manual inspection, 
#so both pdx and ped heatmaps have the same scale (min and max match in both heatmaps)
col_fun= colorRamp2(c(0, 0.25), c("#fef0d9", "#d7301f"))

pdf(paste0(plotpath, "IPD_PDX_QUANTISEQ.pdf"),
   width = 16, height = 10)

heatmap_fx(ipd, pdxes_NS, cohort_ha, pdx_cohort_ha, "QUANTISEQ", col_fun)

dev.off()

# colors set after manual inspection, 
#so both pdx and ped heatmaps have the same scale (min and max match in both heatmaps)
col_fun= colorRamp2(c(0, 0.2), c("#fef0d9", "#d7301f"))

pdf(paste0(plotpath, "IPD_PDX_CIBERSORT.pdf"),
   width = 16, height = 10)

heatmap_fx(ipd, pdxes_NS, cohort_ha, pdx_cohort_ha, "CIBERSORT", col_fun)

dev.off()

col_fun= colorRamp2(c(0, 0.15), c("#fef0d9", "#d7301f"))

pdf(paste0(plotpath, "IPD_PDX_EPIC.pdf"),
   width = 16, height = 10)

heatmap_fx(ipd, pdxes_NS, cohort_ha, pdx_cohort_ha, "EPIC", col_fun)

dev.off()

col_fun= colorRamp2(c(0, 0.6), c("#fef0d9", "#d7301f"))

pdf(paste0(plotpath, "IPD_PDX_TIMER.pdf"),
   width = 16, height = 10)

heatmap_fx(ipd, pdxes_NS, cohort_ha, pdx_cohort_ha, "TIMER", col_fun)

dev.off()

col_fun= colorRamp2(c(0, 60), c("#fef0d9", "#d7301f"))

pdf(paste0(plotpath, "IPD_PDX_MCPCOUNTER.pdf"),
   width = 16, height = 10)

heatmap_fx(ipd, pdxes_NS, cohort_ha, pdx_cohort_ha, "MCPCOUNTER", col_fun)

dev.off()

col_fun= colorRamp2(c(0, 0.3), c("#fef0d9", "#d7301f"))

pdf(paste0(plotpath, "IPD_PDX_XCELL.pdf"),
   width = 16, height = 16)

heatmap_fx(ipd, pdxes_NS, cohort_ha, pdx_cohort_ha, "XCELL", col_fun)

dev.off()
