==> notebooks/20_immune_clusters_tumour_subtypes.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Functions" data-toc-modified-id="Functions-2">Functions</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-3">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-4">Main</a></span></li></ul></div>

# Dependencies

options(scipen = 999)

library(gridExtra)
library(grid)
library(cowplot)

# Functions

source("~/OneDrive - UHN/R_src/ggplot2_theme.R")
source("~/OneDrive - UHN/R_src/Heatmap_functions.R")
source("~/OneDrive - UHN/R_src/plotting_functions.R")
source("~/OneDrive - UHN/R_src/color_schemes.R")

# Paths

datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

load(file = file.path(datapath,"TME_clusters/IPD_estimate_metadata_IC.RData"))

table(metadata_IC$immune_cluster)

table(metadata_IC$cohort)

lggp <- subgroup_IC.fx(metadata_IC, "pedLGG", "Purples")

hggp <- subgroup_IC.fx(metadata_IC, "pedHGG", "Oranges")

atrtp <- subgroup_IC.fx(metadata_IC, "ATRT", "Greens")

epnp <- subgroup_IC.fx(metadata_IC, "EPN", "Reds")

mbp <- subgroup_IC.fx(metadata_IC, "MB", "Blues")

nblp <- subgroup_IC.fx(metadata_IC, "NBL", "GnBu")

p1 <- plot_grid(atrtp+ ggtitle(expression(~underline("ATRT"))),
                epnp + ggtitle(expression(~underline("EPN"))),
                hggp+ ggtitle(expression(~underline("pedHGG"))),
                nblp + ggtitle(expression(~underline("NBL"))), 
                mbp+ ggtitle(expression(~underline("MB"))), 
                lggp+ ggtitle(expression(~underline("pedLGG"))),
                nrow=1 ,
                align="h")

pdf(paste0(plotpath, "IPD_clusters_tumour_subgroups.pdf"),
      width = 50, height = 15, useDingbats = FALSE)

p1

dev.off()
