==> notebooks/30_Mupexi_cleanup.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Dependencies" data-toc-modified-id="Dependencies-1">Dependencies</a></span></li><li><span><a href="#Functions" data-toc-modified-id="Functions-2">Functions</a></span></li><li><span><a href="#Paths" data-toc-modified-id="Paths-3">Paths</a></span></li><li><span><a href="#Main" data-toc-modified-id="Main-4">Main</a></span><ul class="toc-item"><li><span><a href="#CBTTC" data-toc-modified-id="CBTTC-4.1">CBTTC</a></span></li><li><span><a href="#Clean-up-mupexi-files-to-include-mutations-that-are-overlapped-from-strelka-and-mutect2-for-cbttc" data-toc-modified-id="Clean-up-mupexi-files-to-include-mutations-that-are-overlapped-from-strelka-and-mutect2-for-cbttc-4.2">Clean up mupexi files to include mutations that are overlapped from strelka and mutect2 for cbttc</a></span></li><li><span><a href="#TARGET" data-toc-modified-id="TARGET-4.3">TARGET</a></span></li></ul></li></ul></div>

# Dependencies

library(maftools)

# Functions

# Paths

manifestpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Manifests/"
datapath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Data/"
plotpath <- "/Users/anabbi/OneDrive - UHN/Documents/IPD2/Plots/"

# Main

load(file = paste0(datapath,"snv/total_mut_manifest_estimate.RData"))

## CBTTC

CBTTC_mafsummary <- read.table(file = paste0(datapath, "snv/CBTTC_mafsummary_overlapped_3VAF.txt"),
                        sep = "\t", check.names = F,
                        header = TRUE, stringsAsFactors = FALSE)

mymaf_CBTTC <- read.maf(maf =  paste0(datapath, "snv/CBTTC_maf_overlapped_3VAF.maf"),
                      useAll = TRUE, verbose = FALSE,
                      vc_nonSyn = c("Frame_Shift_Del", "Frame_Shift_Ins", 
                                    "Nonsense_Mutation","Nonstop_Mutation", 
                                    "In_Frame_Del","In_Frame_Ins", 
                                    "Missense_Mutation"))

CBTTC_data_maf <- mymaf_CBTTC@data

CBTTC_data_maf$Tumor_Sample_Barcode <- as.character(CBTTC_data_maf$Tumor_Sample_Barcode)

Add sample id to maf file

CBTTC_data_maf_sampleid <- merge(CBTTC_mafsummary[,c("sample_id", "Tumor_Sample_Barcode")], 
                                 CBTTC_data_maf, by = "Tumor_Sample_Barcode")

length(unique(CBTTC_data_maf_sampleid$Tumor_Sample_Barcode))

subset to samples included in total_mut_manifest_estimate_quanti

ls()

CBTTC_data_maf <- CBTTC_data_maf_sampleid[CBTTC_data_maf_sampleid$sample_id %in% total_mut_manifest_estimate$sample_id,]

head(CBTTC_data_maf$sample_id)

length(unique(CBTTC_data_maf$Tumor_Sample_Barcode))

## Clean up mupexi files to include mutations that are overlapped from strelka and mutect2 for cbttc

CBTTC manifest that optitype and mutect2 vcf files are matched with sample ids and all. Use this to map mupexi filenames to sample_id

opti_mutect2_cbttc <- read.table(file = paste0(manifestpath, "mutect2_optitype_manifest_cbttc.txt"),
                        sep = "\t", check.names = F, 
                        header = TRUE, stringsAsFactors = FALSE)

dim(opti_mutect2_cbttc)

head(opti_mutect2_cbttc$name)

datapath

mountpath <- "/Users/anabbi/Desktop/H4H/immpedcan/antigen_prediction/output/"

flelist <- list.files(mountpath, pattern = "mupexi")

length(flelist)

length(unique(CBTTC_data_maf$sample_id))

for(f in flelist){
    print(f)
    #get the filename
    myfilename <- gsub(".mupexi", "", f)
    #get sample_id by grepping filename from manifest file
    mysample_id <- opti_mutect2_cbttc$sample_id[grepl(myfilename, opti_mutect2_cbttc$name)]
    #skip if sample_id is not in CBTTC_data_maf
    if(sum(CBTTC_data_maf$sample_id == mysample_id) == 0){ 
        message("sample_id not found")
        next}
    # get overlapped vars for the sample_id
    myoverlappedmutations <- CBTTC_data_maf[CBTTC_data_maf$sample_id == mysample_id,]
    
    #read the mupexi file
    mymupexi <- read.table(file = paste0(mountpath, f),
                        sep = "\t", check.names = F, comment.char = "#",
                        header = TRUE, stringsAsFactors = FALSE)
    #skip if genes in mupexi output arent in overlapped mutations
    if(sum(mymupexi$Gene_ID %in% myoverlappedmutations$Gene) == 0){ 
        message("Genes in mupexi not found in overlapped variants")
        next}
    #subset to overlapped vars only using Ensembl IDs
    mupexi_genes <- mymupexi[ mymupexi$Gene_ID %in% myoverlappedmutations$Gene,]
    mupexi_genes_AA <- mupexi_genes[ mupexi_genes$Amino_Acid_Change %in% myoverlappedmutations$Amino_acids,]
    # annotate binders for Norm and Mut
    mupexi_genes_AA$Norm_MHCbinder <- NA
    mupexi_genes_AA$Norm_MHCbinder[mupexi_genes_AA$Norm_MHCrank_BA <= 0.5] <- "SB"
    mupexi_genes_AA$Norm_MHCbinder[mupexi_genes_AA$Norm_MHCrank_BA > 0.5 &
                                   mupexi_genes_AA$Norm_MHCrank_BA <= 2] <- "WB"


    mupexi_genes_AA$Mut_MHCbinder <- NA
    mupexi_genes_AA$Mut_MHCbinder[mupexi_genes_AA$Mut_MHCrank_BA <= 0.5] <- "SB"
    mupexi_genes_AA$Mut_MHCbinder[mupexi_genes_AA$Mut_MHCrank_BA > 0.5 &
                                  mupexi_genes_AA$Mut_MHCrank_BA <= 2] <- "WB"
    
      
    #Include all peptides
    mupexi_genes_AA_binders <- mupexi_genes_AA #[ !is.na(mupexi_genes_AA$Mut_MHCbinder),]
    
    #skip if no SB or WB in mupexi output arent in overlapped mutations
    if(nrow(mupexi_genes_AA_binders) == 0){ 
        message("No SB or WB found in Mut_MHC")
        next} 
    # add sample_id as variable
    mupexi_genes_AA_binders$sample_id <- mysample_id
    write.table(mupexi_genes_AA_binders,
           file = paste0(datapath, "Mupexi/", mysample_id, "_overlappedvar_all.mupexi"), sep = "\t", quote = F, row.names = F)
    
}

## TARGET

mymaf_gdc <- read.maf(maf =  paste0(datapath, "snv/GDC_allmaf_3VAF_fl_dedup_sampleid_KB.maf"),
                      useAll = TRUE, verbose = FALSE,
                      vc_nonSyn = c("Frame_Shift_Del", "Frame_Shift_Ins", 
                                    "Nonsense_Mutation","Nonstop_Mutation", 
                                    "In_Frame_Del","In_Frame_Ins", 
                                    "Missense_Mutation"))

mymaf_gdc_data<- mymaf_gdc@data

mymaf_gdc_data$Tumor_Sample_Barcode <- as.character(mymaf_gdc_data$Tumor_Sample_Barcode)

Subset to TARGET only

mymaf_target_data <- mymaf_gdc_data[grepl("TARGET", mymaf_gdc_data$Tumor_Sample_Barcode),]

subset to samples included in total_mut_manifest_estimate_quanti

mymaf_target_data <- mymaf_target_data[mymaf_target_data$Tumor_Sample_Barcode %in% 
                                       total_mut_manifest_estimate$sample_id,]

colnames(mymaf_target_data)[colnames(mymaf_target_data) == "Tumor_Sample_Barcode"] <- "sample_id"

length(unique(mymaf_target_data$sample_id))

read in optitype-mutect2 manifest

opti_mutect2_target <- read.table(file = paste0(manifestpath, "mutect2_optitype_manifest_target.txt"),
                        sep = "\t", check.names = F, 
                        header = TRUE, stringsAsFactors = FALSE)

for(f in flelist){
    print(f)
    #get the filename
    myfilename <- gsub(".mupexi", "", f)
    #get sample_id by grepping filename from manifest file
    mysample_id <- opti_mutect2_target$sample_id.x[grepl(myfilename, opti_mutect2_target$name.y)]
    #skip if sample_id is not in CBTTC_data_maf
    if(sum(mymaf_target_data$sample_id == mysample_id) == 0){ 
        message("sample_id not found")
        next}
    # get overlapped vars for the sample_id
    myoverlappedmutations <- mymaf_target_data[mymaf_target_data$sample_id == mysample_id,]
    
    #read the mupexi file
    mymupexi <- read.table(file = paste0(mountpath, f),
                        sep = "\t", check.names = F, comment.char = "#",
                        header = TRUE, stringsAsFactors = FALSE)
    #skip if genes in mupexi output arent in overlapped mutations
    if(sum(mymupexi$Gene_ID %in% myoverlappedmutations$Gene) == 0){ 
        message("Genes in mupexi not found in overlapped variants")
        next}
    #subset to overlapped vars only using Ensembl IDs
    mupexi_genes <- mymupexi[ mymupexi$Gene_ID %in% myoverlappedmutations$Gene,]
    mupexi_genes_AA <- mupexi_genes[ mupexi_genes$Amino_Acid_Change %in% myoverlappedmutations$Amino_acids,]
    # annotate binders for Norm and Mut
    mupexi_genes_AA$Norm_MHCbinder <- NA
    mupexi_genes_AA$Norm_MHCbinder[mupexi_genes_AA$Norm_MHCrank_BA <= 0.5] <- "SB"
    mupexi_genes_AA$Norm_MHCbinder[mupexi_genes_AA$Norm_MHCrank_BA > 0.5 &
                                   mupexi_genes_AA$Norm_MHCrank_BA <= 2] <- "WB"


    mupexi_genes_AA$Mut_MHCbinder <- NA
    mupexi_genes_AA$Mut_MHCbinder[mupexi_genes_AA$Mut_MHCrank_BA <= 0.5] <- "SB"
    mupexi_genes_AA$Mut_MHCbinder[mupexi_genes_AA$Mut_MHCrank_BA > 0.5 &
                                  mupexi_genes_AA$Mut_MHCrank_BA <= 2] <- "WB"
    
    #Include only those that Mut_MHCbinder is eather WB or SB
    mupexi_genes_AA_binders <- mupexi_genes_AA#[ !is.na(mupexi_genes_AA$Mut_MHCbinder),]
    #skip if no SB or WB in mupexi output arent in overlapped mutations
    if(nrow(mupexi_genes_AA_binders) == 0){ 
        message("No SB or WB found in Mut_MHC")
        next} 
    # add sample_id as variable
    mupexi_genes_AA_binders$sample_id <- mysample_id    
    write.table(mupexi_genes_AA_binders,
           file = paste0(datapath, "Mupexi/", mysample_id, "_overlappedvar_all.mupexi"), sep = "\t", quote = F, row.names = F)
    
}
